// services/digimal/api.ts
export interface GetBreadCrumbInfoResult {
    /**
     * 
     * @type {number}
     * @memberof GetBreadCrumbInfoResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetBreadCrumbInfoResult
     */
    'persianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetBreadCrumbInfoResult
     */
    'englishName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetBreadCrumbInfoResult
     */
    'parentId'?: number | null;
}
// ...
export interface GetProductDetailsResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'productAndPropertyPriceId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'persianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'englishName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'introduction'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'expertCheck'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'brandId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'brandPersianName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'brandEnglishName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'b2CPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'remainingCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'barCode'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductDetailsResult
     */
    'productTypeId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'productPersianTypesName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductDetailsResult
     */
    'productEnglishTypesName'?: string | null;
    /**
     * 
     * @type {GetPropertiesPricesResult}
     * @memberof GetProductDetailsResult
     */
    'pricesInfo'?: GetPropertiesPricesResult;
    /**
     * 
     * @type {Array<GetProductPropertyResult>}
     * @memberof GetProductDetailsResult
     */
    'properties'?: Array<GetProductPropertyResult> | null;
    /**
     * 
     * @type {Array<GetProductSpecificationResult>}
     * @memberof GetProductDetailsResult
     */
    'specifications'?: Array<GetProductSpecificationResult> | null;
    /**
     * 
     * @type {GetProductMediaResult}
     * @memberof GetProductDetailsResult
     */
    'media'?: GetProductMediaResult;
    /**
     * 
     * @type {GetSpecialSaleResult}
     * @memberof GetProductDetailsResult
     */
    'specialSale'?: GetSpecialSaleResult;
    /**
     * 
     * @type {Array<GetResellerInfoResult>}
     * @memberof GetProductDetailsResult
     */
    'resellers'?: Array<GetResellerInfoResult> | null;
    /**
     * 
     * @type {Array<GetBreadCrumbInfoResult>}
     * @memberof GetProductDetailsResult
     */
    'breadCrumbs'?: Array<GetBreadCrumbInfoResult> | null;
}
// ...
export interface GetProductMediaResult {
    /**
     * 
     * @type {Array<MultimediaCapabilities>}
     * @memberof GetProductMediaResult
     */
    'allMedias'?: Array<MultimediaCapabilities> | null;
    /**
     * 
     * @type {number}
     * @memberof GetProductMediaResult
     */
    'productId'?: number;
}
// ...
export interface GetProductPropertyResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductPropertyResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductPropertyResult
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductPropertyResult
     */
    'value'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetProductPropertyResult
     */
    'isFilter'?: boolean;
}
// ...
export interface GetProductSpecificationResult {
    /**
     * 
     * @type {number}
     * @memberof GetProductSpecificationResult
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetProductSpecificationResult
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetProductSpecificationResult
     */
    'value'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetProductSpecificationResult
     */
    'isFilter'?: boolean;
}
// ...
export interface GetPropertiesPricesResult {
    /**
     * 
     * @type {Array<GetPropertyPriceResult>}
     * @memberof GetPropertiesPricesResult
     */
    'propertiesPrices'?: Array<GetPropertyPriceResult> | null;
}
/**
 * 
 * @export
 * @interface GetPropertyPriceResult
 */
export interface GetPropertyPriceResult {
    /**
     * 
     * @type {number}
     * @memberof GetPropertyPriceResult
     */
    'productAndPropertyPriceId'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetPropertyPriceResult
     */
    'b2BPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetPropertyPriceResult
     */
    'b2CPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetPropertyPriceResult
     */
    'productPropertyId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GetPropertyPriceResult
     */
    'warranty'?: string | null;
    /**
     * 
     * @type {SendType}
     * @memberof GetPropertyPriceResult
     */
    'sendTypeId'?: SendType;
    /**
     * 
     * @type {number}
     * @memberof GetPropertyPriceResult
     */
    'remainingCount'?: number;
    /**
     * 
     * @type {GetSpecialSaleInfoResult}
     * @memberof GetPropertyPriceResult
     */
    'specialSale'?: GetSpecialSaleInfoResult;
}
// ...
export interface GetResellerInfoResult {
    /**
     * 
     * @type {number}
     * @memberof GetResellerInfoResult
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetResellerInfoResult
     */
    'productAndPropertyPriceId'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetResellerInfoResult
     */
    'storeName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetResellerInfoResult
     */
    'productPropertyId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GetResellerInfoResult
     */
    'warranty'?: string | null;
    /**
     * 
     * @type {SendType}
     * @memberof GetResellerInfoResult
     */
    'sendTypeId'?: SendType;
    /**
     * 
     * @type {number}
     * @memberof GetResellerInfoResult
     */
    'b2CPrice'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetResellerInfoResult
     */
    'remainingCount'?: number;
    /**
     * 
     * @type {GetSpecialSaleInfoResult}
     * @memberof GetResellerInfoResult
     */
    'specialSale'?: GetSpecialSaleInfoResult;
}
// ...
export interface GetSpecialSaleInfoResult {
    /**
     * 
     * @type {number}
     * @memberof GetSpecialSaleInfoResult
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetSpecialSaleInfoResult
     */
    'specialPrice'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetSpecialSaleInfoResult
     */
    'startSpecialSaleDate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetSpecialSaleInfoResult
     */
    'endSpecialSaleDate'?: string | null;
}
/**
 * 
 * @export
 * @interface GetSpecialSaleResult
 */
export interface GetSpecialSaleResult {
    /**
     * 
     * @type {number}
     * @memberof GetSpecialSaleResult
     */
    'specialSaleId'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof GetSpecialSaleResult
     */
    'specialPrice'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetSpecialSaleResult
     */
    'startSpecialSaleDate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetSpecialSaleResult
     */
    'endSpecialSaleDate'?: string | null;
}
// ...
export interface MultimediaCapabilities {
    /**
     * 
     * @type {number}
     * @memberof MultimediaCapabilities
     */
    'multimediaDetailId'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof MultimediaCapabilities
     */
    'isMain'?: boolean;
    /**
     * 
     * @type {MediaType}
     * @memberof MultimediaCapabilities
     */
    'mediaTypeId'?: MediaType;
    /**
     * 
     * @type {MediaSize}
     * @memberof MultimediaCapabilities
     */
    'mediaSizeId'?: MediaSize;
    /**
     * 
     * @type {string}
     * @memberof MultimediaCapabilities
     */
    'url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MultimediaCapabilities
     */
    'caption'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MultimediaCapabilities
     */
    'fileAddress'?: string | null;
    /**
     * 
     * @type {MultimediaInfo}
     * @memberof MultimediaCapabilities
     */
    'mediaInfo'?: MultimediaInfo;
    /**
     * 
     * @type {MediaCategory}
     * @memberof MultimediaCapabilities
     */
    'mediaCategoryId'?: MediaCategory;
}
// ...
export const SendType = {
    NUMBER_0: 0
} as const;

export type SendType = typeof SendType[keyof typeof SendType];